+++ showhive-boot/admin/build.gradle
@@ -4,7 +4,7 @@ bootJar { enabled = true }
 jar { enabled = false }
 
 dependencies {
-    implementation project(':showhive-data:data')
+    implementation project(':showhive-domain:domain')
     implementation project(':showhive-infra:infra-persistence')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
