+++ showhive-boot/api/src/main/java/com/showhive/auth/api/dto/NaverUserInfo.java
