+++ showhive-boot/api/src/main/java/com/showhive/auth/application/dto/AuthDto.java
