+++ showhive-boot/api/src/main/java/com/showhive/auth/client/KakaoProperties.java
