+++ showhive-boot/api/src/main/java/com/showhive/auth/client/NaverProperties.java
