+++ showhive-boot/api/src/main/java/com/showhive/auth/utils/CookieManager.java
