--- showhive-boot/api/src/main/java/com/showhive/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.showhive.exception;
-
-public record ErrorResponse(
-        String message
-) {
-}
