--- showhive-boot/api/src/main/java/com/showhive/member/application/command/MemberSignUpCommand.java
@@ -1,4 +1,0 @@
-package com.showhive.member.application.command;
-
-public record MemberSignUpCommand(String name) {
-}
